mesh-table: fix broken refresh-routes
refresh-routes was broken in the way that it selected random distances at
which identifiers fall into a bucket for discovery, instead of the actual
identifier. We use the reversible distance metric (logxor) to obtain the
identifier at the distance generated at random. A hack that masked this bug
is removed.